Refactored JTW token create method to public class Refactored AccountController to create & return JWT token on user create
