+++ LoginService/LoginServiceTests/AccountControllerTests.cs
@@ -38,7 +38,7 @@ namespace LoginServiceTests
 
             user = await Context.Users.SingleAsync();
 
-            var target = new AccountController(UserManager);
+            var target = new AccountController(UserManager, ConfigMock.Object);
 
             var claims = new List<Claim>
             {
@@ -72,7 +72,7 @@ namespace LoginServiceTests
         [TestMethod]
         public async Task AccountControllerTestCreate()
         {
-            var target = new AccountController(UserManager);
+            var target = new AccountController(UserManager, ConfigMock.Object);
 
             var request = new ApiUserModel
             {
@@ -103,7 +103,7 @@ namespace LoginServiceTests
         [TestMethod]
         public async Task AccountControllerTestCreateNoPassword()
         {
-            var target = new AccountController(UserManager);
+            var target = new AccountController(UserManager, ConfigMock.Object);
 
             var request = new ApiUserModel
             {
@@ -119,7 +119,7 @@ namespace LoginServiceTests
         [TestMethod]
         public async Task AccountControllerTestCreateNoUsername()
         {
-            var target = new AccountController(UserManager);
+            var target = new AccountController(UserManager, ConfigMock.Object);
 
             var request = new ApiUserModel
             {
@@ -135,7 +135,7 @@ namespace LoginServiceTests
         [TestMethod]
         public async Task AccountControllerTestCreateEmptyUsername()
         {
-            var target = new AccountController(UserManager);
+            var target = new AccountController(UserManager, ConfigMock.Object);
 
             var request = new ApiUserModel
             {
@@ -152,7 +152,7 @@ namespace LoginServiceTests
         [TestMethod]
         public async Task AccountControllerTestCreateNoEmail()
         {
-            var target = new AccountController(UserManager);
+            var target = new AccountController(UserManager, ConfigMock.Object);
 
             var request = new ApiUserModel
             {
@@ -168,7 +168,7 @@ namespace LoginServiceTests
         [TestMethod]
         public async Task AccountControllerTestCreateInvalidEmail()
         {
-            var target = new AccountController(UserManager);
+            var target = new AccountController(UserManager, ConfigMock.Object);
 
             var request = new ApiUserModel
             {
@@ -185,7 +185,7 @@ namespace LoginServiceTests
         [TestMethod]
         public async Task AccountControllerTestCreateEmptyBody()
         {
-            var target = new AccountController(UserManager);
+            var target = new AccountController(UserManager, ConfigMock.Object);
 
             var response = (BadRequestObjectResult)await target.Post(null);
 
@@ -195,7 +195,7 @@ namespace LoginServiceTests
         [TestMethod]
         public async Task AccountControllerTestCreateDuplicate()
         {
-            var target = new AccountController(UserManager);
+            var target = new AccountController(UserManager, ConfigMock.Object);
 
             var request = new ApiUserModel
             {
@@ -230,7 +230,7 @@ namespace LoginServiceTests
             userExist = await Context.Users.AnyAsync(u => u.Id == user2.Id);
             Assert.IsTrue(userExist);
 
-            var target = new AccountController(UserManager);
+            var target = new AccountController(UserManager, ConfigMock.Object);
 
             var result = (OkResult)await target.Delete(user1.Id);
             Assert.AreEqual(200, result.StatusCode);
@@ -259,7 +259,7 @@ namespace LoginServiceTests
 
             user = await Context.Users.SingleAsync();
 
-            var target = new AccountController(UserManager);
+            var target = new AccountController(UserManager, ConfigMock.Object);
 
             var claims = new List<Claim>
             {
@@ -304,7 +304,7 @@ namespace LoginServiceTests
 
             user = await Context.Users.SingleAsync();
 
-            var target = new AccountController(UserManager);
+            var target = new AccountController(UserManager, ConfigMock.Object);
 
             var claims = new List<Claim>
             {
@@ -343,7 +343,7 @@ namespace LoginServiceTests
 
             user = await Context.Users.SingleAsync();
 
-            var target = new AccountController(UserManager);
+            var target = new AccountController(UserManager, ConfigMock.Object);
 
             var claims = new List<Claim>
             {
@@ -381,7 +381,7 @@ namespace LoginServiceTests
 
             user = await Context.Users.SingleAsync();
 
-            var target = new AccountController(UserManager);
+            var target = new AccountController(UserManager, ConfigMock.Object);
 
             var claims = new List<Claim>
             {
@@ -427,7 +427,7 @@ namespace LoginServiceTests
 
             user = await Context.Users.SingleAsync();
 
-            var target = new AccountController(UserManager);
+            var target = new AccountController(UserManager, ConfigMock.Object);
 
             var claims = new List<Claim>
             {
